Swap in timing utils from Notes
