+++ test/provider/expect.unittest.ts
@@ -12,7 +12,7 @@ import {createMockEnvironmnent, createMockAppWindow} from '../mocks';
 import {Application} from '../../src/client/main';
 import {ContextChannel} from '../../src/provider/model/ContextChannel';
 import {AppWindow} from '../../src/provider/model/AppWindow';
-import {advanceTime, useFakeTime} from '../demo/utils/time';
+import {advanceTime, useMockTime} from '../utils/unit/time';
 import {DeferredPromise} from '../../src/provider/common/DeferredPromise';
 import {PartiallyWritable} from '../types';
 import {Timeouts} from '../../src/provider/constants';
@@ -67,7 +67,7 @@ beforeEach(async () => {
     model = new Model(mockAppDirectory, mockEnvironment, mockApiHandler);
 
     jest.resetAllMocks();
-    useFakeTime();
+    useMockTime();
 });
 
 describe('When creating a directory FDC3 app', () => {
